@@ -11,6 +11,7 @@
     "pack": "electron-builder --dir",
     "dist": "electron-builder"
   },
+  "type": "module",
   "repository": "https://github.com/tu4mo/noter",
   "author": "tu4mo",
   "license": "ISC",
@@ -19,7 +20,7 @@
     "lodash.debounce": "^4.0.8"
   },
   "devDependencies": {
-    "electron": "^25.1.0",
+    "electron": "^28.0.0",
     "electron-builder": "^21.2.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^6.15.0",
